py/wbr/requirements.txt
pandas>=2.0.0
gspread>=5.0.0
google-auth>=2.0.0

